move TextScan to com.apple.foundationdb.record.query.plan package
