+++ fdb-record-layer-core/src/com/apple/foundationdb/record/query/plan/planning/TextScanPlanner.java
@@ -35,6 +35,7 @@ import com.apple.foundationdb.record.query.expressions.NestedField;
 import com.apple.foundationdb.record.query.expressions.OneOfThemWithComponent;
 import com.apple.foundationdb.record.query.expressions.QueryComponent;
 import com.apple.foundationdb.record.query.plan.ScanComparisons;
+import com.apple.foundationdb.record.query.plan.TextScan;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
